refac: server모듈 아키텍처 변경
