--- server/src/main/kotlin/kpring/server/application/port/input/AddUserAtServerUseCase.kt
@@ -1,16 +1,0 @@
-package kpring.server.application.port.input
-
-import kpring.core.server.dto.request.AddUserAtServerRequest
-
-interface AddUserAtServerUseCase {
-  fun inviteUser(
-    serverId: String,
-    invitorId: String,
-    userId: String,
-  )
-
-  fun addInvitedUser(
-    serverId: String,
-    req: AddUserAtServerRequest,
-  )
-}
