--- server/src/main/kotlin/kpring/server/application/port/input/CreateServerUseCase.kt
@@ -1,8 +1,0 @@
-package kpring.server.application.port.input
-
-import kpring.core.server.dto.request.CreateServerRequest
-import kpring.core.server.dto.response.CreateServerResponse
-
-interface CreateServerUseCase {
-  fun createServer(req: CreateServerRequest): CreateServerResponse
-}
