--- server/src/main/kotlin/kpring/server/application/port/input/DeleteServerUseCase.kt
@@ -1,13 +1,0 @@
-package kpring.server.application.port.input
-
-interface DeleteServerUseCase {
-  fun deleteServer(
-    serverId: String,
-    userId: String,
-  )
-
-  fun deleteServerMember(
-    serverId: String,
-    userId: String,
-  )
-}
