--- server/src/main/kotlin/kpring/server/application/port/input/GetCategoryUseCase.kt
@@ -1,7 +1,0 @@
-package kpring.server.application.port.input
-
-import kpring.core.server.dto.CategoryInfo
-
-interface GetCategoryUseCase {
-  fun getCategories(): List<CategoryInfo>
-}
