--- server/src/main/kotlin/kpring/server/application/port/input/GetServerInfoUseCase.kt
@@ -1,16 +1,0 @@
-package kpring.server.application.port.input
-
-import kpring.core.server.dto.ServerInfo
-import kpring.core.server.dto.ServerSimpleInfo
-import kpring.core.server.dto.request.GetServerCondition
-
-interface GetServerInfoUseCase {
-  fun getServerInfo(serverId: String): ServerInfo
-
-  fun getServerList(
-    condition: GetServerCondition,
-    userId: String,
-  ): List<ServerSimpleInfo>
-
-  fun getOwnedServerList(userId: String): List<ServerSimpleInfo>
-}
