--- server/src/main/kotlin/kpring/server/application/port/input/UpdateHostAtServerUseCase.kt
@@ -1,11 +1,0 @@
-package kpring.server.application.port.input
-
-import kpring.core.server.dto.request.UpdateHostAtServerRequest
-
-interface UpdateHostAtServerUseCase {
-  fun updateServerHost(
-    serverId: String,
-    userId: String,
-    otherUser: UpdateHostAtServerRequest,
-  )
-}
