--- server/src/main/kotlin/kpring/server/application/port/output/DeleteServerPort.kt
@@ -1,9 +1,0 @@
-package kpring.server.application.port.output
-
-import kpring.server.domain.ServerProfile
-
-interface DeleteServerPort {
-  fun delete(serverId: String)
-
-  fun deleteMember(serverProfile: ServerProfile)
-}
