--- server/src/main/kotlin/kpring/server/application/port/output/GetServerPort.kt
@@ -1,9 +1,0 @@
-package kpring.server.application.port.output
-
-import kpring.server.domain.Server
-
-interface GetServerPort {
-  fun get(id: String): Server
-
-  fun getServerWith(userId: String): List<Server>
-}
