--- server/src/main/kotlin/kpring/server/application/port/output/SaveServerPort.kt
@@ -1,7 +1,0 @@
-package kpring.server.application.port.output
-
-import kpring.server.domain.Server
-
-interface SaveServerPort {
-  fun create(server: Server): Server
-}
