--- server/src/main/kotlin/kpring/server/application/port/output/UpdateServerProfilePort.kt
@@ -1,7 +1,0 @@
-package kpring.server.application.port.output
-
-import kpring.server.domain.ServerProfile
-
-interface UpdateServerProfilePort {
-  fun updateServerHost(serverProfile: ServerProfile)
-}
