+++ server/src/main/kotlin/kpring/server/service/CategoryService.kt
@@ -1,0 +1,7 @@
+package kpring.server.service
+
+import kpring.core.server.dto.CategoryInfo
+
+interface CategoryService {
+  fun getCategories(): List<CategoryInfo>
+}
